Add Bitshares icon to OSX app bundle
Previously the app bundle was created iconless. Now it is created with
the bitshares.icns OSX icon, so it looks much prettier.

Caveats: CMake's bundle creation stuff seems to be broken, so I'm forced
to add a hacky command to copy the icon into the bundle manually, since
CMake isn't doing it like it's supposed to. It works, though.