Remove unecessary cmake guard
Previously CMake was configured to run lineman on every build, which is
very slow, so a guard was placed around it to only build it when
explicitly enabled. I've replaced this guard by making it into a target
which is not depended on by anything else, so that target must be
manually built to cause lineman to run.