Make login flow take user input
So far the login URL handling required the wallet to be unlocked, and
assumed the user wanted to log in as "nathan". Now the user will be
prompted for his password if he is not logged in, and then asked which
account he would like to login with, if more than one account exist.

Also, the password prompt is now a sheet on OSX.