Add license agreement
Now a license agreement is displayed, and the app will not start
until the user has accepted it. When the user accepts it, a value
is set in QSettings so that the dialog is not displayed again.
Any time the app starts and that value is not set correctly, the
agreement will be displayed.