Improve focus siezure when activated by second process
If a second process of the qt_wallet is started, it does not actually
create a window and start a wallet; instead, it notifies the first
process that it should take focus. Until now, this was shaky at best on
OSX, and did not work at all if the window was hidden to the system
tray. Now the window will be unhidden if necessary, and will attempt to
take focus, even on OSX.