Finish minimal viable implementation of dynamic web updates
This basic version allows a server to host a single signed copy of the
web interface, which the client will download upon startup, check the
signature using a hard-coded public key, and use in place of the web UI
built in.

This implementation is basic, probably fragile, and definitely
susceptible to replay attacks.